Changes currency fields to decimal in migrations

+++ database/migrations/2016_03_01_125345_create_account_contributions_migration_table.php
@@ -16,7 +16,7 @@ class CreateAccountContributionsMigrationTable extends Migration
             $table->increments('id');
             $table->integer('account_id')->unsigned();
             $table->integer('user_id')->unsigned();
-            $table->integer('amount')->unsigned();
+            $table->decimal('amount', 15, 2)->unsigned();
             $table->timestamps();
         });
     }
